[Feat] #458 - Move AppDelegate
