+++ HappyAnding/HappyAnding/Views/HappyAndingApp.swift
@@ -7,11 +7,7 @@
 
 import SwiftUI
 
-import Firebase
 import FirebaseCore
-import FirebaseFirestore
-import FirebaseMessaging
-
 
 @main
 struct HappyAndingApp: App {
@@ -101,89 +97,5 @@ struct HappyAndingApp: App {
                 break
             }
         }
-    }
-}
-
-class AppDelegate: NSObject, UIApplicationDelegate {
-    
-    let gcmMessageIDKey = "gcm.message_id"
-    
-    // APN 등록 및 사용자 권한 받기
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-        
-        // 원격 알림 등록
-        UNUserNotificationCenter.current().delegate = self
-        
-        let authOption: UNAuthorizationOptions = [.alert, .badge, .sound]
-        UNUserNotificationCenter.current().requestAuthorization(
-            options: authOption,
-            completionHandler: { _, _ in }
-        )
-        
-        application.registerForRemoteNotifications()
-        
-        // 메시지 대리자 설정
-        Messaging.messaging().delegate = self
-        
-        return true
-    }
-    
-    // APNs 토큰과 등록 토큰 매핑
-    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
-        Messaging.messaging().apnsToken = deviceToken
-    }
-    
-    // 주제 메시지 수신 및 처리
-    func application(_ application: UIApplication,
-                     didReceiveRemoteNotification userInfo: [AnyHashable: Any],
-                     fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
-        
-        Messaging.messaging().appDidReceiveMessage(userInfo)
-        
-        completionHandler(UIBackgroundFetchResult.newData)
-    }
-}
-
-// 토큰 갱신 모니터링
-extension AppDelegate: MessagingDelegate {
-    
-    func messaging(_ messaging: Messaging, didReceiveRegistrationToken fcmToken: String?) {
-        
-        let dataDict: [String: String] = ["token": fcmToken ?? ""]
-        NotificationCenter.default.post(
-            name: Notification.Name("FCMToken"),
-            object: nil,
-            userInfo: dataDict
-        )
-        
-        // TODO: If necessary send token to application server.
-        
-    }
-}
-
-// 알림 처리
-extension AppDelegate: UNUserNotificationCenterDelegate {
-    
-    func userNotificationCenter(_ center: UNUserNotificationCenter,
-                                willPresent notification: UNNotification,
-                                withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
-        
-        let userInfo = notification.request.content.userInfo
-        
-        // 알림 수신 정보를 애널리틱스에 전달
-        Messaging.messaging().appDidReceiveMessage(userInfo)
-                
-        completionHandler([[.banner, .badge, .sound]])
-    }
-    
-    func userNotificationCenter(_ center: UNUserNotificationCenter,
-                                didReceive response: UNNotificationResponse,
-                                withCompletionHandler completionHandler: @escaping () -> Void) {
-        
-        let userInfo = response.notification.request.content.userInfo
-        
-        Messaging.messaging().appDidReceiveMessage(userInfo)
-        
-        completionHandler()
     }
 }
